refactor: move source plan repository to persistence/jooq structure
